+++ R/class-workbook-wrappers.R
@@ -354,7 +354,8 @@ wb_add_data <- function(
 #' It is possible to leave the cell empty `"none"` or to create a text input
 #' using a named character with name `text` like: `c(text = "Total")`.
 #' It's also possible to pass other spreadsheet software functions if they
-#' return a single value and hence `"SUM"` would work too.
+#' return a single value and hence `"SUM"` would work too. Its also possible
+#' to provide a custom formula body.
 #'
 #' @family worksheet content functions
 #' @family workbook wrappers